applied recommended changes
